requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
cairocffi==0.8.0
CairoSVG==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
dateparser==0.7.0
defusedxml==0.5.0
Django==2.0.5
django-crispy-forms==1.7.2
django-macros==0.4.0
django-qr-code==1.0.0
gunicorn==19.8.1
html5lib==1.0.1
humanize==0.5.1
idna==2.6
lxml==4.2.1
maya==0.5.0
mysqlclient==1.3.12
pdfrw==0.4
pendulum==1.5.1
Pillow==5.1.0
pycparser==2.18
Pyphen==0.9.4
python-dateutil==2.7.3
pytz==2018.4
pytzdata==2018.5
qrcode==6.0
regex==2018.6.6
requests==2.18.4
six==1.11.0
snaptime==0.2.4
tinycss2==0.6.1
toolz==0.9.0
tzlocal==1.5.1
urllib3==1.22
WeasyPrint==0.42.3
webencodings==0.5.1
